pyproject.toml
[tool.poetry]
name = "indexor"
version = "0.1.0"
description = ""
authors = ["michael1011 <me@michael1011.at>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-bitcoinrpc = "^1.0"
ruff = "^0.0.265"
psycopg2 = "^2.9.6"
tomli = "^2.0.1"

[tool.poetry.scripts]
indexor = "indexor.indexor:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["ALL"]
ignore = [
    "T201", "D101", "D211", "D213", "INP001", "BLE001", "FBT001", "FBT002", "FBT003",
    "D104", "D100", "D103", "ANN206", "D102", "D107", "ANN101"
]
